requirements.txt
Pillow
numpy
pickle-mixin
ffmpeg-python
matplotlib

